# contracts/openapi/layer4-agents.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Layer 4: Agentic Workflow Engine",
    "description": "LangGraph-powered workflow orchestration for Value Fabric with multi-agent support",
    "version": "0.2.0"
  },
  "paths": {
...
    "/v1/accounts": {
      "get": {
        "tags": [
          "Accounts",
          "Accounts"
        ],
        "summary": "List Accounts",
        "description": "List accounts with filtering and pagination.",
        "operationId": "list_accounts_v1_accounts_get",
        "parameters": [
          {
            "name": "provider",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CRMProvider"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Provider"
            }
          },
          {
            "name": "stage",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Stage"
            }
          },
          {
            "name": "industry",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Industry"
            }
          },
          {
            "name": "region",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Region"
            }
          },
          {
            "name": "segment",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Segment"
            }
          },
          {
            "name": "owner_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Owner Id"
            }
          },
          {
            "name": "sync_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/SyncStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sync Status"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(name|updated_at|company_size|last_synced_at)$",
              "default": "updated_at",
              "title": "Sort By"
            }
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(asc|desc)$",
              "default": "desc",
              "title": "Sort Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "layer4-agents"
        ],
        "summary": "POST /v1/accounts",
        "description": "Discovered from router source (value-fabric\\layer4-agents\\src\\api\\routes\\accounts.py).",
        "operationId": "discovered_post_v1_accounts",
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error"
          }
        }
      }
    },
    "/v1/accounts/search": {
      "post": {
        "tags": [
          "Accounts",
          "Accounts"
        ],
        "summary": "Search Accounts",
        "description": "Search accounts across name, domain, and owner.",
        "operationId": "search_accounts_v1_accounts_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/accounts/sync": {
      "post": {
        "tags": [
          "Accounts",
          "Accounts"
        ],
        "summary": "Sync Accounts",
        "description": "Trigger manual sync for accounts.",
        "operationId": "sync_accounts_v1_accounts_sync_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncAccountsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncAccountsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/accounts/{account_id}": {
      "get": {
        "tags": [
          "Accounts",
          "Accounts"
        ],
        "summary": "Get Account",
        "description": "Get full account detail.",
        "operationId": "get_account_v1_accounts__account_id__get",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountDetailSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/accounts/{account_id}/activity": {
      "get": {
        "tags": [
          "Accounts",
          "Accounts"
        ],
        "summary": "Get Account Activity",
        "description": "Get account activity timeline.",
        "operationId": "get_account_activity_v1_accounts__account_id__activity_get",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Account Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "since_days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 365,
              "minimum": 1,
              "default": 90,
              "title": "Since Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountActivityResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/accounts/{account_id}/refresh": {
      "post": {
        "tags": [
          "Accounts",
          "Accounts"
        ],
        "summary": "Refresh Account",
        "description": "Refresh account data from CRM provider.",
        "operationId": "refresh_account_v1_accounts__account_id__refresh_post",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountDetailSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/hypotheses/account/{account_id}": {
      "get": {
        "tags": [
          "layer4-agents"
        ],
        "summary": "GET /v1/hypotheses/account/{account_id}",
        "description": "Discovered from router source (value-fabric\\layer4-agents\\src\\api\\routes\\value_hypotheses.py).",
        "operationId": "discovered_get_v1_hypotheses_account_account_id",
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AccountActivityResponse": {
        "properties": {
          "account_id": {
            "type": "string",
            "format": "uuid",
            "title": "Account Id"
          },
          "interactions": {
            "items": {
              "$ref": "#/components/schemas/ActivityItemSchema"
            },
            "type": "array",
            "title": "Interactions"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          }
        },
        "type": "object",
        "required": [
          "account_id",
          "interactions",
          "total_count"
        ],
        "title": "AccountActivityResponse",
        "description": "Account activity timeline response."
      },
...
      "AccountListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AccountListItemSchema"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "has_more"
        ],
        "title": "AccountListResponse",
        "description": "Paginated account list response."
      },
      "AccountSearchRequest": {
        "properties": {
          "query": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query",
            "description": "Search across name, domain, owner"
          },
          "provider": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CRMProvider"
              },
              {
                "type": "null"
              }
            ]
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage"
          },
          "industry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industry"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "segment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Segment"
          },
          "owner_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Id"
          },
          "sync_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SyncStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page",
            "default": 1
          },
          "page_size": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Page Size",
            "default": 20
          },
          "sort_by": {
            "type": "string",
            "pattern": "^(name|updated_at|company_size|last_synced_at)$",
            "title": "Sort By",
            "default": "updated_at"
          },
          "sort_order": {
            "type": "string",
            "pattern": "^(asc|desc)$",
            "title": "Sort Order",
            "default": "desc"
          }
        },
        "type": "object",
        "title": "AccountSearchRequest",
        "description": "Account search request body."
      },
...
      "CRMProvider": {
        "type": "string",
        "enum": [
          "salesforce",
          "hubspot"
        ],
        "title": "CRMProvider",
        "description": "Supported CRM providers."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SyncAccountsRequest": {
        "properties": {
          "provider": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CRMProvider"
              },
              {
                "type": "null"
              }
            ],
            "description": "Specific provider to sync, or null for all"
          },
          "account_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Ids",
            "description": "Specific account IDs to sync, or null for all"
          },
          "force_refresh": {
            "type": "boolean",
            "title": "Force Refresh",
            "description": "Force re-sync even if data is fresh",
            "default": false
          }
        },
        "type": "object",
        "title": "SyncAccountsRequest",
        "description": "Manual sync trigger request."
      },
      "SyncAccountsResponse": {
        "properties": {
          "sync_id": {
            "type": "string",
            "title": "Sync Id",
            "description": "Unique sync job identifier"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Sync status: queued, running, completed, failed"
          },
          "provider": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CRMProvider"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "sync_id",
          "status",
          "message"
        ],
        "title": "SyncAccountsResponse",
        "description": "Sync trigger response."
      },
      "SyncStatus": {
        "type": "string",
        "enum": [
          "synced",
          "pending",
          "failed",
          "stale"
        ],
        "title": "SyncStatus",
        "description": "Account sync status from CRM provider."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}